pyproject.toml
[tool.poetry]
name = "netobs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pysnmplib = "^5.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
